Extract meetings-management IPC handlers to src/ipc/
main.ts has grown to 2,476 lines with 46 IPC handlers, module state, and
lifecycle code interleaved. This first modularization PR introduces an
IpcContext-based pattern and moves the meetings-management cluster
(delete-meeting, export-recording-m4a, merge-recordings) out of main.ts.

Changes:
- New src/ipc/types.ts defining IpcContext (service getters, broadcast
  helpers, mainWindow accessor). Service getters are functions because
  config changes recreate service instances.
- New src/ipc/register.ts exposing registerAllIpc(ctx) as the single
  entry point; subsequent IPC modules will register here.
- New src/ipc/meetingsManagement.ts containing the three handlers and
  their module-local mergeInFlight flag. Dependencies are pulled
  through ctx or imported directly from leaf modules.
- src/main.ts: ~300 lines removed (handlers + helper), single
  registerAllIpc(buildContext()) call added. No behavior change.

Verification:
- pnpm run build: pass
- pnpm run lint: 0 warnings, 0 errors
- pnpm run format:check: pass
- pnpm test: 320/320 pass
